Added real-time progress bar

+++ webhooks/orders/order-paid.php
@@ -4,11 +4,21 @@
 use WPS\DB\Settings_Connection;
 use WPS\Cart;
 use WPS\Transients;
+use WPS\Webhooks;
+use WPS\WS;
 
-$Connection = new Settings_Connection();
+$jsonData = file_get_contents('php://input');
 
-$order = json_decode( file_get_contents('php://input') );
 
-$cartID = Cart::wps_get_cart_id_from_order($order);
+if (Webhooks::webhook_verified($jsonData, WS::get_header_hmac())) {
 
-Transients::delete('wps_cart_' . $cartID);
+  $order = json_decode($jsonData);
+  $cartID = Cart::wps_get_cart_id_from_order($order);
+
+  Transients::delete('wps_cart_' . $cartID);
+
+  do_action('wps_webhook_checkouts_order_paid', $order);
+
+} else {
+  error_log('WP Shopify Error - Unable to verify response from order-paid webhook');
+}
